simple implementation of polling and event_invoking on fd

+++ src/lib/commonallocator.h
@@ -8,7 +8,7 @@
 #ifndef COMMONALLOCATOR_H_
 #define COMMONALLOCATOR_H_
 
-void *tc_malloc(size_t len);
+void *tc_malloc(int len);
 void tc_free(void *ptr);
 
 #define NEW(type, obj, size) \
